// src/interfaces/iIP.ts
export interface iIP {
  asn: string;
  city: string;
  continent_code: string;
  country: string;
  country_code: string;
  country_area: string;
  country_calling_code: string;
  country_capital: string;
  country_code_iso3: string;
  country_name: string;
  country_population: number;
  country_tld: string;
  currency: string;
  currency_name: string;
  in_eu: false;
  ip: string;
  languages: string;
  latitude: number;
  longitude: number;
  org: string;
  postal: string;
  region: string;
  region_code: string;
  timezone: string;
  utc_offset: string;
  version: string;
}